feat/ footer como componente
