+++ src/app/components/dashboard/dashboard-gastos/dashboard-gastos.html
@@ -2,8 +2,7 @@
     <app-navbar></app-navbar>
 
     <main class="flex-grow w-full max-w-[1600px] mx-auto p-6 lg:p-8 pt-20">
-        
-        <!-- Header Dashboard -->
+
         <div class="flex flex-col md:flex-row justify-between items-center mb-10 gap-6">
             <div class="text-left w-full md:w-auto">
                 <div class="flex items-center gap-3">
@@ -19,7 +18,6 @@
                 </p>
             </div>
 
-            <!-- Navegación -->
             <div class="flex gap-3 bg-white p-2 rounded-2xl shadow-sm border border-slate-200">
                 <button
                     (click)="navegarA('/home')"
@@ -42,7 +40,6 @@
             </div>
         </div>
 
-        <!-- Cargando -->
         <div *ngIf="cargando" class="text-center py-20 opacity-60">
             <div class="animate-spin rounded-full h-12 w-12 border-b-4 border-red-500 mx-auto mb-4"></div>
             <p class="text-slate-500 text-xl font-medium">Procesando datos...</p>
@@ -50,10 +47,8 @@
 
         <div *ngIf="!cargando" class="space-y-8 animate-in fade-in duration-500">
             
-            <!-- Tarjetas Stats -->
             <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
-                
-                <!-- Total Gastos -->
+
                 <div class="bg-white p-6 rounded-2xl shadow-lg border border-slate-100 hover:shadow-xl transition duration-300">
                     <div class="flex justify-between items-start mb-2">
                         <h3 class="text-xs font-bold text-slate-400 uppercase tracking-wider">Total Gastos</h3>
@@ -64,7 +59,6 @@
                     </p>
                 </div>
 
-                <!-- Promedio -->
                 <div class="bg-white p-6 rounded-2xl shadow-lg border border-slate-100 hover:shadow-xl transition duration-300">
                     <div class="flex justify-between items-start mb-2">
                         <h3 class="text-xs font-bold text-slate-400 uppercase tracking-wider">Promedio / Gasto</h3>
@@ -75,7 +69,6 @@
                     </p>
                 </div>
 
-                <!-- Cantidad -->
                 <div class="bg-white p-6 rounded-2xl shadow-lg border border-slate-100 hover:shadow-xl transition duration-300">
                     <div class="flex justify-between items-start mb-2">
                         <h3 class="text-xs font-bold text-slate-400 uppercase tracking-wider">N° Transacciones</h3>
@@ -86,7 +79,6 @@
                     </p>
                 </div>
 
-                <!-- Mayor Gasto -->
                 <div class="bg-white p-6 rounded-2xl shadow-lg border border-slate-100 hover:shadow-xl transition duration-300 relative overflow-hidden group">
                     <div class="absolute right-0 top-0 h-full w-1 bg-gradient-to-b from-red-400 to-red-600"></div>
                     <h3 class="text-xs font-bold text-slate-400 uppercase tracking-wider mb-2">Mayor Gasto</h3>
@@ -100,7 +92,6 @@
                 </div>
             </div>
 
-            <!-- Gráficas -->
             <div class="grid grid-cols-1 lg:grid-cols-2 gap-8">
                 <div class="bg-white p-6 rounded-2xl shadow-xl border border-slate-100 flex flex-col h-[450px]">
                     <h3 class="text-lg font-bold text-slate-700 mb-6 text-center border-b border-slate-100 pb-3">
@@ -121,7 +112,6 @@
                 </div>
             </div>
 
-            <!-- Top Gastos -->
             <div class="bg-white p-6 rounded-2xl shadow-xl border border-slate-100">
                 <div class="flex items-center gap-2 mb-6 border-b border-slate-100 pb-3">
                     <span class="text-2xl">⚠️</span>
