+++ src/app/components/footer/footer.html
@@ -1,0 +1,3 @@
+<footer class="p-4 bg-orange-50 backdrop-blur-sm text-center text-sm text-slate-900 border-t border-white/5">
+© 2025 MiGasto - Todos los derechos reservados
+</footer>
