+++ src/app/components/formularios/profile-form/profile-form.html
@@ -1,0 +1,82 @@
+<section class="min-h-screen flex flex-col bg-stone-100 text-white relative">
+    <app-navbar></app-navbar>
+
+    <main class="flex-grow flex flex-col items-center p-6 pt-24">
+        <h2 class="text-4xl font-extrabold mb-8 text-sky-950/80">Editar Perfil</h2>
+
+        <form
+            [formGroup]="form"
+            (ngSubmit)="onSubmit()"
+            class="bg-white p-8 rounded-3xl shadow-xl w-full max-w-md space-y-5 border border-white/60"
+        >
+            <div>
+                <label class="block text-xs font-bold text-slate-500 uppercase mb-1 ml-1">Nombre completo</label>
+                <input
+                    formControlName="nombre_completo"
+                    type="text"
+                    class="w-full px-4 py-3 rounded-xl bg-slate-50 border border-slate-200 focus:outline-none focus:ring-2 focus:ring-sky-500 text-slate-700 font-medium transition"
+                />
+            </div>
+
+            <div>
+                <label class="block text-xs font-bold text-slate-500 uppercase mb-1 ml-1">Edad</label>
+                <input
+                    formControlName="edad"
+                    type="number"
+                    class="w-full px-4 py-3 rounded-xl bg-slate-50 border border-slate-200 focus:outline-none focus:ring-2 focus:ring-sky-500 text-slate-700 font-medium transition"
+                />
+            </div>
+
+            <div>
+                <label class="block text-xs font-bold text-slate-500 uppercase mb-1 ml-1">Foto de perfil (URL)</label>
+                <input
+                    formControlName="foto_perfil"
+                    type="text"
+                    class="w-full px-4 py-3 rounded-xl bg-slate-50 border border-slate-200 focus:outline-none focus:ring-2 focus:ring-sky-500 text-slate-700 font-medium transition placeholder-slate-300"
+                    placeholder="https://..."
+                />
+            </div>
+
+            <div>
+                <label class="block text-xs font-bold text-slate-500 uppercase mb-1 ml-1">Teléfono</label>
+                <input
+                    formControlName="telefono"
+                    type="text"
+                    class="w-full px-4 py-3 rounded-xl bg-slate-50 border border-slate-200 focus:outline-none focus:ring-2 focus:ring-sky-500 text-slate-700 font-medium transition"
+                />
+            </div>
+
+            <div>
+                <label class="block text-xs font-bold text-slate-500 uppercase mb-1 ml-1">Género</label>
+                <select
+                    formControlName="id_genero"
+                    class="w-full px-4 py-3 rounded-xl bg-slate-50 border border-slate-200 focus:outline-none focus:ring-2 focus:ring-sky-500 text-slate-700 font-medium transition cursor-pointer"
+                >
+                    <option value="" class="text-slate-400">Seleccionar</option>
+                    <option [ngValue]="1">Masculino</option>
+                    <option [ngValue]="2">Femenino</option>
+                    <option [ngValue]="3">Otro</option>
+                    <option [ngValue]="4">Prefiero no decirlo</option>
+                </select>
+            </div>
+
+            <div class="flex justify-center gap-4 pt-6">
+                <button
+                    type="button"
+                    routerLink="/profile"
+                    class="px-6 py-2.5 rounded-full font-bold text-slate-500 hover:bg-slate-100 transition border border-transparent hover:border-slate-200"
+                >
+                    Cancelar
+                </button>
+                <button
+                    type="submit"
+                    class="bg-sky-600 hover:bg-sky-700 text-white font-bold px-8 py-2.5 rounded-full shadow-lg shadow-sky-200 transition transform hover:-translate-y-0.5"
+                >
+                    Guardar Cambios
+                </button>
+            </div>
+        </form>
+    </main>
+
+    <app-footer></app-footer>
+</section>
