+++ src/app/pages/auth/reset-password/reset-password.html
@@ -1,4 +1,4 @@
-<div class="min-h-screen flex items-center justify-center bg-gradient-to-b from-sky-200 to-sky-950 p-4 font-sans">
+<div class="min-h-screen flex items-center justify-center bg-stone-100 p-4 font-sans">
     <div class="bg-white p-8 rounded-xl shadow-lg w-full max-w-md border border-gray-200">
