--- src/app/pages/gastos/gastos-form/gastos-form.html
@@ -1,88 +1,0 @@
-<form (ngSubmit)="guardarGasto()" class="flex flex-col space-y-5">
-
-  <div>
-    <label class="block text-xs font-bold text-slate-600 uppercase mb-1 ml-1">Descripción</label>
-    <input
-      [(ngModel)]="gasto.descripcion"
-      name="descripcion"
-      type="text"
-      class="w-full p-3 rounded-xl bg-slate-50 border border-slate-200 text-slate-700 placeholder-slate-400 outline-none focus:ring-2 focus:ring-red-500 transition"
-      placeholder="Ej. Transporte, Comida..."
-      required
-    />
-  </div>
-
-  <div>
-    <label class="block text-xs font-bold text-slate-600 uppercase mb-1 ml-1">Monto</label>
-    <div class="relative">
-        <span class="absolute left-3 top-3 text-slate-400 font-bold">$</span>
-        <input
-          [(ngModel)]="gasto.monto"
-          name="monto"
-          type="number"
-          class="w-full p-3 pl-8 rounded-xl bg-slate-50 border border-slate-200 text-slate-700 placeholder-slate-400 outline-none focus:ring-2 focus:ring-red-500 transition font-bold"
-          placeholder="0.00"
-          required
-        />
-    </div>
-  </div>
-
-  <div>
-    <label class="block text-xs font-bold text-slate-600 uppercase mb-1 ml-1">Fecha</label>
-    <input
-      [(ngModel)]="gasto.fecha"
-      name="fecha"
-      type="date"
-      class="w-full p-3 rounded-xl bg-slate-50 border border-slate-200 text-slate-700 outline-none focus:ring-2 focus:ring-red-500 transition"
-    />
-  </div>
-
-  <div>
-    <label class="block text-xs font-bold text-slate-600 uppercase mb-1 ml-1">Moneda</label>
-    <select
-      [(ngModel)]="gasto.id_moneda"
-      name="id_moneda"
-      class="w-full p-3 rounded-xl bg-slate-50 border border-slate-200 text-slate-700 outline-none focus:ring-2 focus:ring-red-500 transition cursor-pointer"
-    >
-      <option [ngValue]="null" class="text-slate-400">Seleccionar...</option>
-      <option *ngFor="let moneda of monedas" [ngValue]="moneda.id_moneda">
-        {{ moneda.codigo }} {{ moneda.nombre }}
-      </option>
-    </select>
-  </div>
-
-  <div>
-    <label class="block text-xs font-bold text-slate-600 uppercase mb-1 ml-1">Tags</label>
-    <select
-      multiple
-      [(ngModel)]="gasto.tags"
-      name="tags"
-      class="w-full p-3 rounded-xl bg-slate-50 border border-slate-200 text-slate-700 outline-none focus:ring-2 focus:ring-red-500 transition h-32"
-    >
-      <option *ngFor="let tag of tagsDisponibles" [ngValue]="tag.id_tag" class="p-2 hover:bg-red-50 rounded mb-1 cursor-pointer">
-        {{ tag.nombre }}
-      </option>
-    </select>
-    <p class="text-xs text-slate-400 mt-2 ml-1 italic flex items-center gap-1">
-      ℹ Mantén presionado Ctrl (o Cmd) para seleccionar varios.
-    </p>
-  </div>
-
-  <div class="flex justify-end space-x-3 pt-4 border-t border-slate-100 mt-2">
-    <button
-      type="button"
-      class="px-5 py-2.5 rounded-xl font-bold text-slate-500 hover:bg-slate-100 transition border border-transparent hover:border-slate-200"
-      (click)="cancelarEdicion()"
-    >
-      Cancelar
-    </button>
-
-    <button
-      type="submit"
-      class="px-6 py-2.5 rounded-xl font-bold text-white shadow-lg transition transform active:scale-95 flex items-center gap-2"
-      [ngClass]="gasto.id_movimiento ? 'bg-amber-500 hover:bg-amber-600 shadow-amber-200' : 'bg-red-600 hover:bg-red-700 shadow-red-200'"
-    >
-      {{ gasto.id_movimiento ? 'Actualizar Gasto' : 'Registrar Gasto' }}
-    </button>
-  </div>
-</form>
