+++ src/app/pages/home/home.html
@@ -90,7 +90,5 @@
         </div>
     </main>
 
-    <footer class="p-4 bg-stone-200 text-center text-sm text-slate-600 border-t border-white/10">
-        © 2025 MiGasto - Todos los derechos reservados
-    </footer>
+    <app-footer></app-footer>
 </section>
