+++ src/app/pages/home/home.ts
@@ -2,13 +2,14 @@ import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { Footer } from '../../components/footer/footer';
 import { Navbar } from '../../components/navbar/navbar';
 import { BalanceService, BalanceSummary } from '../../services/balance.service';
 
 @Component({
   selector: 'app-home',
   standalone: true,
-  imports: [CommonModule, RouterModule, Navbar, HttpClientModule], 
+  imports: [CommonModule, RouterModule, Navbar, HttpClientModule, Footer], 
   templateUrl: './home.html',
 })
 export class Home implements OnInit {
