+++ src/app/pages/prestamos/prestamos.html
@@ -9,8 +9,6 @@
         </div>
 
         <div class="flex flex-col lg:flex-row gap-8 items-start">
-            
-            <!-- 1. FORMULARIO (Crear Préstamo) -->
             <div class="w-full lg:w-1/3 bg-white/90 backdrop-blur-md rounded-2xl shadow-2xl p-6 border border-white/50 lg:sticky lg:top-24 z-10">
                 <div class="mb-6 border-l-4 border-sky-500 pl-4">
                     <h2 class="text-2xl font-bold text-slate-800">Nueva Deuda</h2>
@@ -18,15 +16,12 @@
                 </div>
 
                 <form [formGroup]="prestamoForm" (ngSubmit)="guardarPrestamo()" class="flex flex-col space-y-4">
-                    
-                    <!-- Prestamista -->
                     <div>
                         <label class="block text-xs font-bold text-slate-600 uppercase mb-1 ml-1">Prestamista / Banco</label>
                         <input type="text" formControlName="prestamista" placeholder="Ej: Banco Z" 
                             class="w-full bg-slate-50 border border-slate-200 rounded-lg p-3 outline-none focus:ring-2 focus:ring-sky-500 text-slate-700 transition">
                     </div>
 
-                    <!-- Monto -->
                     <div>
                         <label class="block text-xs font-bold text-slate-600 uppercase mb-1 ml-1">Monto Total</label>
                         <div class="relative">
@@ -36,14 +31,12 @@
                         </div>
                     </div>
 
-                    <!-- Fecha -->
                     <div>
                         <label class="block text-xs font-bold text-slate-600 uppercase mb-1 ml-1">Fecha Límite</label>
                         <input type="date" formControlName="fecha_limite" 
                             class="w-full bg-slate-50 border border-slate-200 rounded-lg p-3 outline-none focus:ring-2 focus:ring-sky-500 text-slate-700 transition">
                     </div>
 
-                    <!-- Descripción -->
                     <div>
                         <label class="block text-xs font-bold text-slate-600 uppercase mb-1 ml-1">Descripción</label>
                         <textarea formControlName="descripcion" rows="2" placeholder="Notas..." class="w-full bg-slate-50 border border-slate-200 rounded-lg p-3 outline-none focus:ring-2 focus:ring-sky-500 text-slate-700 resize-none"></textarea>
@@ -56,19 +49,15 @@
                 </form>
             </div>
 
-            <!-- 2. LISTA DE PRÉSTAMOS -->
             <div class="w-full lg:w-2/3 flex flex-col space-y-4">
                 <h2 class="text-xl font-semibold text-slate-800 mb-2 pl-2 border-l-4 border-emerald-400">Historial</h2>
 
-                <!-- Cargando -->
                 <div *ngIf="loading && prestamos.length === 0" class="text-center text-white p-4">Cargando préstamos...</div>
 
-                <!-- Lista Vacía -->
                 <div *ngIf="!loading && prestamos.length === 0" class="bg-white/80 backdrop-blur rounded-xl p-8 text-center shadow-lg">
                     <p class="text-slate-500 text-lg">No tienes préstamos registrados.</p>
                 </div>
 
-                <!-- Tarjeta de Préstamo -->
                 <div *ngFor="let p of prestamos" class="bg-white rounded-xl p-5 shadow-lg flex flex-col justify-between hover:shadow-xl transition duration-300 border-l-4"
                     [ngClass]="p.estado === 'pagado' ? 'border-emerald-500 opacity-80' : 'border-red-500'">
                     
@@ -83,7 +72,6 @@
                         </div>
                     </div>
 
-                    <!-- Barra de Progreso -->
                     <div class="mb-4">
                         <div class="flex justify-between text-xs font-bold mb-1">
                             <span class="text-emerald-600">Pagado: ${{ p.monto_pagado | number }}</span>
@@ -97,25 +85,19 @@
                     </div>
 
                     <div class="pt-3 mt-4 border-t border-slate-100">
-    
-    <!-- Fecha Límite (La dejamos arriba para que no estorbe en móviles) -->
-    <span *ngIf="p.fecha_limite" class="block mb-2 text-xs text-slate-400 flex items-center gap-1">
-        📅 Límite: {{ p.fecha_limite | date:'mediumDate' }}
-    </span>
 
-    <!-- CONTENEDOR FLEX PRINCIPAL: Separa los extremos -->
+                        <span *ngIf="p.fecha_limite" class="block mb-2 text-xs text-slate-400 flex items-center gap-1">
+                            📅 Límite: {{ p.fecha_limite | date:'mediumDate' }}
+                        </span>
+
                         <div class="flex items-center justify-between">
-                            
-                            <!-- 👈 IZQUIERDA: Botón Detalles -->
                             <button (click)="abrirHistorial(p.id_prestamo!)" 
                                 class="text-xs font-bold text-slate-400 hover:text-sky-600 flex items-center gap-1 transition group">
                                 <span class="group-hover:underline">Ver detalles de abono</span>
                             </button>
 
-                            <!-- 👉 DERECHA: Grupo de Acciones (Eliminar + Abonar) -->
                             <div class="flex items-center gap-2">
-                                
-                                <!-- Botón Eliminar -->
+
                                 <button *ngIf="p.id_prestamo" 
                                     (click)="eliminarPrestamo(p.id_prestamo!)" 
                                     class="p-2 text-slate-300 hover:text-red-500 hover:bg-red-50 rounded-lg transition"
@@ -125,7 +107,6 @@
                                     </svg>
                                 </button>
 
-                                <!-- Botón Abonar -->
                                 <button *ngIf="p.estado !== 'pagado'" 
                                     (click)="abrirModalAbono(p)" 
                                     class="px-4 py-2 bg-emerald-100 text-emerald-700 hover:bg-emerald-600 hover:text-white rounded-lg text-sm font-bold transition flex items-center gap-1 shadow-sm">
@@ -134,14 +115,12 @@
                                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 4v16m8-8H4" />
                                     </svg>
                                 </button>
-                                
-                                <!-- Badge Pagado -->
+                        
                                 <span *ngIf="p.estado === 'pagado'" 
                                     class="px-3 py-1 bg-slate-100 text-slate-500 rounded-lg text-xs font-bold border border-slate-200">
                                     ✅ Pagado
                                 </span>
                             </div>
-
                         </div>
                     </div>
                 </div>
@@ -149,7 +128,6 @@
         </div>
     </main>
 
-    <!-- Modal de Abono -->
     <app-abono-modal 
         *ngIf="prestamoSeleccionado" 
         [prestamo]="prestamoSeleccionado"
@@ -164,8 +142,6 @@
         (cambioRealizado)="onHistorialChange()">
     </app-historial-abonos-modal>
 
-    <footer class="p-4 bg-stone-400 backdrop-blur-sm text-center text-sm text-slate-900 border-t border-white/5">
-    © 2025 MiGasto - Todos los derechos reservados
-    </footer>
+    <app-footer></app-footer>
 
 </section>
