+++ src/app/services/auth.service.ts
@@ -18,7 +18,6 @@ export class AuthService {
 
   constructor(private http: HttpClient) {}
 
-  // --- MÉTODOS EXISTENTES ---
 
   register(data: { email: string; password: string; rolId: number }): Observable<any> {
     return this.http.post(`${this.API_URL}/register`, data);
@@ -50,7 +49,7 @@ export class AuthService {
       
       return Date.now() < decoded.exp * 1000;
     } catch {
-      return false; // Token malformado
+      return false; 
     }
   }
 
@@ -73,18 +72,10 @@ export class AuthService {
     localStorage.removeItem(this.TOKEN_KEY);
   }
 
-  // --- NUEVOS MÉTODOS PARA RESET PASSWORD ---
-
-  /**
-   * Envía el correo al backend para solicitar el código OTP.
-   */
   solicitarRecuperacion(email: string): Observable<any> {
     return this.http.post(`${this.API_URL}/request-password-reset`, { email });
   }
 
-  /**
-   * Envía el código OTP recibido y la nueva contraseña para actualizarla.
-   */
   restablecerContrasena(otp: string, newPassword: string): Observable<any> {
     return this.http.post(`${this.API_URL}/reset-password`, { otp, newPassword });
   }
